services/payment-core-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic-settings==2.5.0
python-dotenv==1.0.1
aiokafka==0.11.0
celery[redis]==5.4.0
redis==5.1.0
kafka-python-ng==2.2.2
uuid6==2025.0.1
httpx==0.27.0

structlog==24.4.0
prometheus-client==0.20.0
